Made add_review screen

+++ domain/src/main/java/com/appninjas/domain/model/Review.kt
@@ -1,3 +1,3 @@
 package com.appninjas.domain.model
 
-data class Review(val reviewerName: String, val reviewRate: Int, val reviewDate: String, val reviewText: String)
+data class Review(val reviewerName: String, val reviewRate: Float, val reviewDate: String, val reviewText: String)
